+++ src/incache.js
@@ -2,15 +2,19 @@ const helper = require('./helper');
 const fs = require('fs');
 
 class InCache {
+
     /**
      * Create instance
      * @param [opts] {Object} configuration object
+     * @param [opts.maxAge=0] {number} max age in milliseconds. If 0 not expire
+     * @param [opts.expires] {Date|string} a Date for expiration. (overwrites `opts.maxAge`)
+     * @param [opts.silent=false] {boolean} if true no event will be triggered
      * @param [opts.save=true] {boolean} if true saves cache in disk. (server only)
      * @param [opts.filePath=.incache] {string} cache file path
      * @param [opts.storeName] {string} store name
-     * @param [opts.global] {Object} global record configuration
-     * @param [opts.global.silent=false] {boolean} if true no event will be triggered
-     * @param [opts.global.life=0] {number} max age. If 0 not expire
+     * @param [opts.global] {Object} **deprecated:** global record configuration
+     * @param [opts.global.silent=false] {boolean} **deprecated:** if true no event will be triggered, use `silent` instead
+     * @param [opts.global.life=0] {number} **deprecated:** max age in seconds. If 0 not expire, use `maxAge` instead
      * @constructor
      */
     constructor(opts = {}) {
@@ -38,6 +42,8 @@ class InCache {
             storeName: '',
             save: true,
             filePath: '.incache',
+            maxAge: 0,
+            silent: false,
             global: {
                 silent: false,
                 life: 0
@@ -85,12 +91,7 @@ class InCache {
     /**
      * Set configuration
      * @param [opts] {Object} configuration object
-     * @param [opts.save=true] {boolean} if true saves cache in disk. (server only)
-     * @param [opts.filePath=.incache] {string} cache file path
-     * @param [opts.storeName] {string} store name
-     * @param [opts.global] {Object} global record configuration
-     * @param [opts.global.silent=false] {boolean} if true no event will be triggered
-     * @param [opts.global.life=0] {number} max age. If 0 not expire
+     * @see {@link constructor} for further information
      */
     setConfig(opts = {}) {
         if (opts.storeName)
@@ -104,6 +105,11 @@ class InCache {
                 data: {},
                 config: this.DEFAULT_CONFIG
             };
+        }
+
+        if(opts.global) {
+            helper.deprecated(opts.global.life, 'global.life is deprecated use maxAge instead');
+            helper.deprecated(opts.global.silent, 'global.silent is deprecated use silent instead');
         }
 
         this._root[this.GLOBAL_KEY].config = helper.defaults(opts, this.DEFAULT_CONFIG);
@@ -125,17 +131,29 @@ class InCache {
     }
 
     /**
+     * InCache Record
+     * @typedef {Object} InCache~record
+     * @property {boolean} isNew - indicates if is a new record
+     * @property {Date|null} createdOn - creation date
+     * @property {Date|null} updatedOn - update date
+     * @property {Date|null} expiresOn - expiry date
+     * @property {any} value - record value
+     */
+
+    /**
      * Set/update record
      * @param key {any}
      * @param value {any}
      * @param [opts] {Object} options object
      * @param [opts.silent=false] {boolean} if true no event will be triggered. (overwrites global configuration)
-     * @param [opts.life=0] {number} max age. If 0 not expire. (overwrites global configuration)
-     * @returns {{isNew: boolean, createdOn: Date|null, updatedOn: Date|null, value: *}}
+     * @param [opts.maxAge=0] {number} max age in milliseconds. If 0 not expire. (overwrites global configuration)
+     * @param [opts.expires] {Date|string} a Date for expiration. (overwrites global configuration and `opts.maxAge`)
+     * @param [opts.life=0] {number} **deprecated:** max age in seconds. If 0 not expire. (overwrites global configuration)
+     * @returns {InCache~record}
      * @example
      * inCache.set('my key', 'my value');
      * inCache.set('my object', {a: 1, b: 2});
-     * inCache.set('my boolean', true, {life: 2}); // Expires after 2 seconds
+     * inCache.set('my boolean', true, {maxAge: 2000}); // Expires after 2 seconds
      */
     set(key, value, opts = {}) {
         let record = {
@@ -146,10 +164,15 @@ class InCache {
             value: value
         };
 
-        opts = helper.defaults(opts, this.DEFAULT_CONFIG.global);
+        opts = helper.defaults(opts, this.DEFAULT_CONFIG);
 
-        if (opts.life && helper.is(opts.life, 'number')) {
+        if (opts.maxAge && helper.is(opts.maxAge, 'number')) {
+            record.expiresOn = helper.addMSToNow(opts.maxAge);
+        } else if (opts.life && helper.is(opts.life, 'number')) {
+            helper.deprecated(opts.life, 'life is deprecated use maxAge instead');
             record.expiresOn = helper.addSecondsToNow(opts.life);
+        } else if (opts.expires && (helper.is(opts.expires, 'date') || helper.is(opts.expires, 'string'))) {
+            record.expiresOn = new Date(opts.expires);
         }
 
         if (this.has(key)) {
@@ -194,7 +217,7 @@ class InCache {
      * Get record by key
      * @param key {any}
      * @param [onlyValue=true] {boolean} if false get InCache record
-     * @returns {any|null}
+     * @returns {any|null|InCache~record}
      * @example
      * inCache.get('my key');
      */
@@ -228,7 +251,7 @@ class InCache {
      * Given a key that has value like an array adds value to end of array
      * @param key {any}
      * @param value {any}
-     * @returns {*}
+     * @returns {InCache~record}
      * @example
      * inCache.set('myArray', ['hello', 'world']);
      * inCache.addTo('myArray', 'ciao'); //-> ['hello', 'world', 'ciao'];
@@ -249,7 +272,7 @@ class InCache {
      * Given a key that has value like an array adds value to beginning of array
      * @param key {any}
      * @param value {any}
-     * @returns {*}
+     * @returns {InCache~record}
      * @example
      * inCache.set('myArray', ['hello', 'world']);
      * inCache.prependTo('myArray', 'ciao'); //-> ['ciao', 'hello', 'world'];
@@ -485,7 +508,7 @@ class InCache {
      * onCreated callback
      * @callback InCache~createdCallback
      * @param key {string} key of record created
-     * @param record {Object} record object
+     * @param record {InCache~record} record object
      */
 
     /**
@@ -504,7 +527,7 @@ class InCache {
      * onUpdated callback
      * @callback InCache~updatedCallback
      * @param key {string} key of record updated
-     * @param record {Object} record object
+     * @param record {InCache~record} record object
      */
 }