+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.0.0] - 2017-08-22
+- Changed configuration
+    - Added new properties
+        - `maxAge`
+        - `expires`
+        - `silent`
+    - **Deprecated properties**
+        - `global.life`    
+        - `global.silent`    
+
 ## [3.1.1] - 2017-08-21
 - Fixed readme